Refactor DiInfo class to add a static Create method
This commit refactors the DiInfo class by adding a new static Create method. This method allows creating a new instance of DiInfo with the provided data and optional prototypes. If an instance with the same data already exists, an error will be thrown. The new instance is then added to the data map.

Co-authored-by: [Author Name]